- Merged remote-tracking branch 'origin/develop' into MAGETWO-21596

Conflicts:
	app/code/Magento/Backend/Block/Context.php
	app/code/Magento/Backend/Model/Locale/Manager.php
	app/code/Magento/Backend/Model/Translate.php
	app/code/Magento/Backend/etc/adminhtml/events.xml
	app/code/Magento/Checkout/Helper/Data.php
	app/code/Magento/Contacts/Controller/Index.php
	app/code/Magento/Core/Helper/Js.php
	app/code/Magento/Core/Model/App/Area.php
	app/code/Magento/DesignEditor/Helper/Data.php
	app/code/Magento/DesignEditor/Model/Translate/Inline.php
	app/code/Magento/MultipleWishlist/Controller/Index.php
	app/code/Magento/Sales/Controller/Adminhtml/Order.php
	app/code/Magento/Translation/Model/Inline/Config.php
	app/code/Magento/Translation/Model/Inline/Parser.php
	app/code/Magento/Translation/Model/Resource/String.php
	app/code/Magento/VersionsCms/Controller/Adminhtml/Cms/Page/Revision.php
	app/code/Magento/Wishlist/Controller/Index.php
	dev/tests/integration/testsuite/Magento/Core/Helper/JsTest.php
	dev/tests/integration/testsuite/Magento/Data/Argument/Interpreter/StringTest.php
	dev/tests/integration/testsuite/Magento/Email/Model/Template/FilterTest.php
	dev/tests/integration/testsuite/Magento/Translate/InlineTest.php
	dev/tests/integration/testsuite/Magento/TranslateTest.php
	dev/tests/integration/testsuite/Magento/Translation/Controller/AjaxTest.php
	dev/tests/integration/testsuite/Magento/Translation/Model/InlineParserTest.php
	dev/tests/integration/testsuite/Magento/Translation/Model/StringTest.php
	dev/tests/integration/testsuite/Magento/Translation/_files/db_translate.php
	dev/tests/integration/testsuite/Magento/Translation/_files/db_translate_admin_store.php
	dev/tests/static/testsuite/Magento/Test/Legacy/_files/obsolete_classes.php
	dev/tests/static/testsuite/Magento/Test/Legacy/_files/obsolete_methods.php
	dev/tests/static/testsuite/Magento/Test/Legacy/_files/obsolete_properties.php
	dev/tests/unit/testsuite/Magento/Backend/Model/Translate/Inline/ConfigFactoryTest.php
	dev/tests/unit/testsuite/Magento/Backend/Model/TranslateTest.php
	dev/tests/unit/testsuite/Magento/Core/Model/Translate/Inline/ConfigTest.php
	dev/tests/unit/testsuite/Magento/DesignEditor/Helper/DataTest.php
	dev/tests/unit/testsuite/Magento/GiftRegistry/Model/EntityTest.php
	dev/tests/unit/testsuite/Magento/ImportExport/Model/Import/Entity/CustomerCompositeTest.php
	dev/tests/unit/testsuite/Magento/Phrase/Renderer/CompositeTest.php
	dev/tests/unit/testsuite/Magento/Phrase/Renderer/TranslateTest.php
	dev/tests/unit/testsuite/Magento/PhraseTest.php
	dev/tests/unit/testsuite/Magento/Sales/Model/Order/Pdf/AbstractTest.php
	dev/tests/unit/testsuite/Magento/Translate/Inline/ConfigFactoryTest.php
	lib/Magento/Translate.php
	lib/Magento/Translate/Inline.php
	lib/Magento/View/Element/AbstractBlock.php
	lib/Magento/View/Element/Context.php

+++ app/code/Magento/CustomerBalance/Block/Sales/Order/Customerbalance.php
@@ -19,10 +19,8 @@ class Customerbalance extends \Magento\View\Element\Template
      * @param \Magento\View\Element\Template\Context $context
      * @param array $data
      */
-    public function __construct(
-        \Magento\View\Element\Template\Context $context,
-        array $data = array()
-    ) {
+    public function __construct(\Magento\View\Element\Template\Context $context, array $data = array())
+    {
         parent::__construct($context, $data);
         $this->_isScopePrivate = true;
     }
@@ -52,14 +50,16 @@ class Customerbalance extends \Magento\View\Element\Template
      */
     public function initTotals()
     {
-        if ((float)$this->getSource()->getCustomerBalanceAmount() == 0) {
+        if ((double)$this->getSource()->getCustomerBalanceAmount() == 0) {
             return $this;
         }
-        $total = new \Magento\Object(array(
-            'code'      => $this->getNameInLayout(),
-            'block_name'=> $this->getNameInLayout(),
-            'area'      => $this->getArea()
-        ));
+        $total = new \Magento\Object(
+            array(
+                'code' => $this->getNameInLayout(),
+                'block_name' => $this->getNameInLayout(),
+                'area' => $this->getArea()
+            )
+        );
         $after = $this->getAfterTotal();
         if (!$after) {
             $after = 'giftcards';
